Refactor variables processing
